Changed neutralization injection counter from type INTEGER to type REAL as it now add up the (possibly different) MPFs of each particle.

+++ src/particles/emission/particle_emission_vars.f90
@@ -114,9 +114,9 @@ LOGICAL              :: UseNeutralization           ! Flag for counting the char
 CHARACTER(255)       :: NeutralizationSource        ! Name of the boundary for calculating the particle balance
 INTEGER              :: nNeutralizationElems        ! Number of elements used for neutralization source (if required)
 LOGICAL, ALLOCATABLE :: isNeutralizationElem(:)     ! Flag each element if it is a neutralization element
-INTEGER, ALLOCATABLE :: NeutralizationBalanceElem(:)! Number of particles to be emitted within each neutralization element
-INTEGER              :: NeutralizationBalance       ! Counter for charged particles (processor local): Add +1 for electrons and -1 for ions
-INTEGER              :: NeutralizationBalanceGlobal ! Counter for charged particles (global): Add +1 for electrons and -1 for ions
+REAL, ALLOCATABLE    :: NeutralizationBalanceElem(:)! Number of particles to be emitted within each neutralization element
+REAL                 :: NeutralizationBalance       ! Counter for charged particles (processor local): Add +1 for electrons and -1 for ions
+REAL                 :: NeutralizationBalanceGlobal ! Counter for charged particles (global): Add +1 for electrons and -1 for ions
 REAL                 :: NeutralizationBalanceCurrent ! Electric current (positive) for emitted electrons. Note that the sampling time is only considered in the when the data is written to .csv
 
 ! Bulk electron temperature
